Replace end turn penalty with more sophisticated approach
Instead of penalizing end turn every time, penalize end turn only if we have more then half of the map. This makes the AI more aggressive with a higher percentage of map occupancy. This was updated because the AI stopped attacking towards the end of the game and even though the game was very well played, the AI ended up losing due to not attacking.